requirements.txt
httpx
requests
PyPDF2
reportlab
python-dotenv
openai